+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/intrastructure/web/controller/IssuanceV2IT.java
@@ -304,6 +304,7 @@ class IssuanceV2IT {
         credentialManagement = credentialManagementRepository.save(CredentialManagement.builder()
                 .id(UUID.randomUUID())
                 .accessToken(UUID.randomUUID())
+                .credentialManagementStatus(CredentialStatusManagementType.INIT)
                 .accessTokenExpirationTimestamp(Instant.now().plusSeconds(120).getEpochSecond())
                 .renewalRequestCnt(0)
                 .renewalResponseCnt(0)