+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/intrastructure/web/controller/KeyAttestationFlowIT.java
@@ -201,6 +201,7 @@ class KeyAttestationFlowIT {
         var credentialManagement = credentialManagementRepository.save(CredentialManagement.builder()
                 .id(UUID.randomUUID())
                 .accessToken(UUID.randomUUID())
+                .credentialManagementStatus(CredentialStatusManagementType.INIT)
                 .accessTokenExpirationTimestamp(Instant.now().plusSeconds(120).getEpochSecond())
                 .renewalRequestCnt(0)
                 .renewalResponseCnt(0)