+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/service/CredentialManagementServiceIT.java
@@ -83,6 +83,7 @@ class CredentialManagementServiceIT {
         var credentialManagement = credentialManagementRepository.save(CredentialManagement.builder()
                 .id(UUID.randomUUID())
                 .accessToken(UUID.randomUUID())
+                .credentialManagementStatus(CredentialStatusManagementType.INIT)
                 .accessTokenExpirationTimestamp(Instant.now().plusSeconds(120).getEpochSecond())
                 .renewalRequestCnt(0)
                 .renewalResponseCnt(0)