+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/domain/credentialoffer/CredentialStatusManagementType.java
@@ -1,6 +1,7 @@
 package ch.admin.bj.swiyu.issuer.domain.credentialoffer;
 
 public enum CredentialStatusManagementType {
+    INIT,
     ISSUED,
     SUSPENDED,
     REVOKED